backend/pyproject.toml
[project]
name = "vertiche-backend"
version = "1.0.0"
description = "Vertiche Contract Intelligence Platform - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "supabase>=2.7.0",
    "boto3>=1.35.0",
    "anthropic>=0.40.0",
    "sentence-transformers>=3.3.0",
    "chromadb>=0.5.0",
    "langchain-text-splitters>=0.2.4",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "PyJWT[crypto]>=2.8.0",
    "numpy>=1.26.0",
]

[project.scripts]
dev = "uvicorn app.main:app --reload --port 8000"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
